requirements.txt
pymongo~=4.3.3

mysql-connector-python~=8.0.32
Faker~=18.3.1
python-dateutil~=2.8.2

psycopg~=3.1.8
